Rename widget folder to ui
